Change skin loading and support
- Move skin specification to k9s cluster config section
- Load skins for skins dir
