+++ .golangci.yml
@@ -1,10 +1,10 @@
 # options for analysis running
 run:
   # default concurrency is a available CPU number
-  concurrency: 4
+  concurrency: 8
 
   # timeout for analysis, e.g. 30s, 5m, default is 1m
-  timeout: 1m
+  timeout: 5m
 
   # exit code when at least one issue was found, default is 1
   issues-exit-code: 1
